+++ src/examples/zkapps/escrow/escrow.ts
@@ -1,0 +1,38 @@
+import {
+  SmartContract,
+  state,
+  State,
+  method,
+  DeployArgs,
+  Permissions,
+  UInt64,
+  AccountUpdate,
+  PrivateKey,
+  PublicKey,
+} from 'snarkyjs';
+
+export class Escrow extends SmartContract {
+  deploy(args: DeployArgs) {
+    super.deploy(args);
+    this.setPermissions({
+      ...Permissions.default(),
+      editState: Permissions.proof(),
+      send: Permissions.proof(),
+    });
+  }
+
+  @method deposit(user: PublicKey) {
+    // add your deposit logic circuit here
+    // that will adjust the amount
+
+    const payerUpdate = AccountUpdate.create(user);
+    payerUpdate.send({ to: this.address, amount: UInt64.from(1000000) });
+  }
+
+  @method withdraw(user: PublicKey) {
+    // add your withdrawal logic circuit here
+    // that will adjust the amount
+
+    this.send({ to: user, amount: UInt64.from(1000000) });
+  }
+}